// src/frontend/src/backend.d.ts
export interface Client {
    clientId: bigint;
    gstNumber: string;
    name: string;
    email: string;
    address: string;
    companyName: string;
    phone: string;
}